Modernized build system by removing webpack and using parcel
HMR works when running just the client bundler, but the server bundler doesn't work 100% because the server itself reloads when detecting changes. Granted this only happens when a file is modified that's shared by both the server and the client, but once the server shuts off the client won't pick it back up when it automatically comes back on.

Solution is to probably run server and client watchers concurrently but separately and then proxy all requests to the server.